Move tower upgrades scripts to new namespace
