+++ Assets/Prefabs/Towers/slowTower.prefab
@@ -194,7 +194,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 556194159825708486}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fc70776cb92db2c4383c5374c4dfa65f, type: 3}
+  m_Script: {fileID: 11500000, guid: a54fc91e238f9c84da5326a68f010409, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Name: Slower Tower
@@ -607,7 +607,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 4943128654439657085}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fc70776cb92db2c4383c5374c4dfa65f, type: 3}
+  m_Script: {fileID: 11500000, guid: a54fc91e238f9c84da5326a68f010409, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Name: Paralyse Tower
@@ -927,7 +927,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 7889544214916897720}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fc70776cb92db2c4383c5374c4dfa65f, type: 3}
+  m_Script: {fileID: 11500000, guid: a54fc91e238f9c84da5326a68f010409, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Name: Slow Tower
@@ -1240,7 +1240,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8278936702669702381}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fe1d3ce55a05e7945a5e8415183ecae8, type: 3}
+  m_Script: {fileID: 11500000, guid: 769f988b7723e0847a0fca2a1b34eec2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   RootNode: {fileID: 6339331099896013953}
