+++ Assets/Scripts/Towers/TowerBehaviour.cs
@@ -5,6 +5,7 @@ using TDDemo.Assets.Scripts.Enemies;
 using TDDemo.Assets.Scripts.Experience;
 using TDDemo.Assets.Scripts.Extensions;
 using TDDemo.Assets.Scripts.Towers.Actions;
+using TDDemo.Assets.Scripts.Towers.Upgrades;
 using TDDemo.Assets.Scripts.Util;
 using UnityEngine;
 using UnityEngine.Events;
