+++ Assets/Scripts/Towers/Upgrades/UpgradeNode.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fc70776cb92db2c4383c5374c4dfa65f
+guid: a54fc91e238f9c84da5326a68f010409
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
