+++ Assets/Scripts/Towers/Upgrades/UpgradePath.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f849fb5e97eb7dc4e8dfad86f5c5c139
+guid: decd69f2a7f0ae14d85b1a6317060598
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
