+++ Assets/Scripts/UI/UpgradeImage.cs
@@ -1,4 +1,4 @@
-using TDDemo.Assets.Scripts.Towers;
+using TDDemo.Assets.Scripts.Towers.Upgrades;
 using UnityEngine;
 using UnityEngine.UI;
